fix: enforce native-to-target level scaling rules and fix pre-stage query
- Fix SQL syntax error in PreStageDungeonLoot JOIN clause (cr.id = ct.entry)
- Fix item scaling not applying by enabling dynamic on-demand fast-store variant creation
- Enforce level scaling rules for Fixed (H) and Dynamic modes
- Enforce native-to-target level matching rule: drops whose native level matches target level (80 to 80, 70 to 70, 60 to 60) bypass scaling to preserve Blizzard stock stats
- Allow lower-tier drops to properly scale up to higher-level players (e.g. Molten Core level 60 items scale to 80 for level 80 players)
- Update ExcludedLevels default to empty and document behavior
- Update README.md with level scaling rules and updated configuration table

+++ src/ItemScalingBaseline.cpp
@@ -278,7 +278,7 @@ uint16 ItemScalingBaseline::CalculateTargetItemLevel(ItemTemplate const* basePro
     }
     if (l0 == 0)
     {
-        l0 = std::clamp<uint8>(static_cast<uint8>(i0), 1, 80);
+        l0 = static_cast<uint8>(std::clamp<uint32>(i0, 1, 80));
     }
 
     uint8 lTarget = std::clamp<uint8>(targetLevel, 1, MAX_BASELINE_LEVEL);
